requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --extra=dev setup.cfg
#
astroid==3.0.1
    # via pylint
attrs==23.1.0
    # via
    #   libdoc2testbench (setup.cfg)
    #   xsdata-attrs
black==23.11.0
    # via libdoc2testbench (setup.cfg)
build==1.0.3
    # via
    #   check-manifest
    #   pip-tools
certifi==2023.11.17
    # via requests
charset-normalizer==3.3.2
    # via
    #   docformatter
    #   requests
check-manifest==0.49
    # via libdoc2testbench (setup.cfg)
click==8.1.7
    # via
    #   black
    #   click-default-group
    #   pip-tools
    #   xsdata
click-default-group==1.2.4
    # via xsdata
colorama==0.4.6
    # via
    #   build
    #   click
    #   pylint
    #   pytest
coverage[toml]==7.3.2
    # via
    #   coverage
    #   pytest-cov
dill==0.3.7
    # via pylint
docformatter==1.7.5
    # via xsdata
docutils==0.20.1
    # via readme-renderer
exceptiongroup==1.2.0
    # via pytest
flake8==6.1.0
    # via libdoc2testbench (setup.cfg)
idna==3.6
    # via requests
importlib-metadata==6.9.0
    # via
    #   build
    #   keyring
    #   twine
importlib-resources==6.1.1
    # via keyring
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via pylint
jaraco-classes==3.3.0
    # via keyring
jinja2==3.1.2
    # via xsdata
keyring==24.3.0
    # via twine
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via jinja2
mccabe==0.7.0
    # via
    #   flake8
    #   pylint
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.1.0
    # via jaraco-classes
mypy==1.7.1
    # via libdoc2testbench (setup.cfg)
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
nh3==0.2.14
    # via readme-renderer
packaging==23.2
    # via
    #   black
    #   build
    #   pytest
pathspec==0.11.2
    # via black
pip-tools==7.3.0
    # via libdoc2testbench (setup.cfg)
pkginfo==1.9.6
    # via twine
platformdirs==4.0.0
    # via
    #   black
    #   pylint
pluggy==1.3.0
    # via pytest
pycodestyle==2.11.1
    # via flake8
pyflakes==3.1.0
    # via flake8
pygments==2.17.2
    # via
    #   readme-renderer
    #   rich
pylint==3.0.2
    # via libdoc2testbench (setup.cfg)
pyproject-hooks==1.0.0
    # via build
pytest==7.4.3
    # via
    #   libdoc2testbench (setup.cfg)
    #   pytest-cov
pytest-cov==4.1.0
    # via libdoc2testbench (setup.cfg)
pytest-spec==3.2.0
    # via libdoc2testbench (setup.cfg)
pywin32-ctypes==0.2.2
    # via keyring
readme-renderer==42.0
    # via twine
requests==2.31.0
    # via
    #   requests-toolbelt
    #   twine
requests-toolbelt==1.0.0
    # via twine
rfc3986==2.0.0
    # via twine
rich==13.7.0
    # via twine
robotframework==6.1.1
    # via libdoc2testbench (setup.cfg)
ruff==0.1.6
    # via libdoc2testbench (setup.cfg)
six==1.16.0
    # via pytest-spec
tomli==2.0.1
    # via
    #   black
    #   build
    #   check-manifest
    #   coverage
    #   mypy
    #   pip-tools
    #   pylint
    #   pyproject-hooks
    #   pytest
tomlkit==0.12.3
    # via pylint
toposort==1.10
    # via xsdata
twine==4.0.2
    # via libdoc2testbench (setup.cfg)
typing-extensions==4.7.1
    # via
    #   astroid
    #   black
    #   mypy
    #   pylint
    #   rich
    #   xsdata
untokenize==0.1.1
    # via docformatter
urllib3==2.1.0
    # via
    #   requests
    #   twine
wheel==0.42.0
    # via
    #   libdoc2testbench (setup.cfg)
    #   pip-tools
xsdata[cli]==23.8
    # via xsdata-attrs
xsdata-attrs[cli]==23.8
    # via libdoc2testbench (setup.cfg)
zipp==3.17.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
